Create separate Task and GitRemote classes.
